fix: harden request parsing and file handling
- Block path traversal and unsafe multipart upload filenames
- Reject malformed empty header names without crashing
- Validate Content-Encoding for chunked request bodies
- Generate unpredictable session IDs and validate session cookie filenames
- Fix large-file partial-send handling and directory listing links
- Tighten location matching, autoindex config validation, and CGI env handling
- Add regression_security.sh for traversal, malformed headers, chunked encoding, listing links, and session collisions

Verified with:
- make
- bash request_tests/regression_security.sh

+++ inc/ParseRequest.hpp
@@ -10,6 +10,10 @@
 #include <cstring>
 #include <zlib.h>
 #include <ctime>
+#include <cerrno>
+#include <climits>
+#include <cstdlib>
+#include <cctype>
 // #include <brotli/decode.h>
 #include "Server.hpp"
 #include "Config.hpp"
@@ -81,7 +85,7 @@ class ParseRequest{
         std::string                                         HttpProtocolVersion;
         Server                                              *S;
         bool                                                chunkedEncoding;
-        int                                                 contentLength;
+        size_t                                              contentLength;
         int                                                 ContentEncodingType;
         std::vector<std::pair<std::string, std::string> >   Headers;
         bool                                                hasValidHost;
@@ -174,7 +178,7 @@ class ParseRequest{
         const std::vector<std::string>&                     getMatchedLocationAllowedMethods();
         std::vector<std::string>&                           getMultipartBuferBody();
         // std::vector<std::string >	                  		getMultipartBuferBody();
-        int                                                 getMatchedLocationBodySizeMax();
+        size_t                                              getMatchedLocationBodySizeMax();
         Server*                                             getBlockServer();
         std::string                                         getCookie(const std::string& name) const;
         const std::map<std::string, std::string>&           getCookies() const;
